Move blog short URL handling to separate controller
